requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
bitsandbytes==0.49.2
certifi==2026.2.25
charset-normalizer==3.4.4
chess==1.11.2
-e git+https://github.com/LeonSavi/chess_exam.git@87e3037c0aac61afe13990edfa5ed66ebd9974f5#egg=chess_tournament
click==8.3.1
colorlog==6.10.1
comm==0.2.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
datasets==4.6.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.0
executing==2.2.1
filelock==3.24.3
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
optuna==4.7.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
Pygments==2.19.2
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.2.19
requests==2.32.4
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smolagents==1.22.0
SQLAlchemy==2.0.47
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.22.0
